WIP move multiprocess_setup_method from composer to composer-builder
